unable to delete User or TaskStatus if it has at least one active linked Task

+++ app/src/main/java/hexlet/code/app/task/service/TaskStatusService.java
@@ -1,11 +1,13 @@
 package hexlet.code.app.task.service;
 
 import hexlet.code.app.exception.ResourceNotFoundException;
+import hexlet.code.app.exception.UnableDeletionException;
 import hexlet.code.app.task.dto.TaskStatusCreateDTO;
 import hexlet.code.app.task.dto.TaskStatusDTO;
 import hexlet.code.app.task.dto.TaskStatusUpdateDTO;
 import hexlet.code.app.task.mapper.TaskStatusMapper;
 import hexlet.code.app.task.model.TaskStatus;
+import hexlet.code.app.task.repository.TaskRepository;
 import hexlet.code.app.task.repository.TaskStatusRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -18,6 +20,9 @@ import java.util.List;
 public class TaskStatusService {
     @Autowired
     private TaskStatusRepository repository;
+
+    @Autowired
+    private TaskRepository taskRepository;
 
     @Autowired
     private TaskStatusMapper mapper;
@@ -50,6 +55,10 @@ public class TaskStatusService {
 
     @Transactional
     public void delete(Long id) {
+        TaskStatus status = repository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Not found"));
+        if (!taskRepository.findAllByTaskStatus(status).isEmpty()) {
+            throw new UnableDeletionException("Status has active tasks");
+        }
         repository.deleteById(id);
     }
 }
